+++ spec/units/instance_types_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require_relative '../lib/types/error_types'
-require_relative 'fixtures/low_instance'
+require_relative '../../lib/types/error_types'
+require_relative '../fixtures/low_instance'
 
 RSpec.describe 'InstanceTypes' do
   describe '#type_reader' do