+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-Minor features that don't break backwards compatibiliy are released as patches.
+Minor features that don't break backwards compatibility are released as patches.
 
 ## 1.2.0 [UNRELEASED]
 
@@ -19,7 +19,7 @@ Minor features that don't break backwards compatibiliy are released as patches.
 ### Added
 
 - Deep type checking
-- Array subtype expressions
+- Array subtype expressions `Array[String | nil]`
 
 ## 1.0.8